// FPLDraftV2/ClientApp/src/app/FANCAST/models/fancast_draft_squad.ts
import { FancastPlayer } from "./fancast-player";

export class FancastDraftSquad {
  player_1: FancastPlayer;
  player_2: FancastPlayer;
  player_3: FancastPlayer;
  player_4: FancastPlayer;
  player_5: FancastPlayer;
  player_6: FancastPlayer;
  player_7: FancastPlayer;
  player_8: FancastPlayer;
  player_9: FancastPlayer;
  player_10: FancastPlayer;
  player_11: FancastPlayer;
}